fixed ticket_watcher.user_id datatype, moved from BIGINT to INTEGER

+++ scripts/database/otrs-schema.maxdb.sql
@@ -694,7 +694,7 @@ CREATE INDEX index_time_accountin45 ON time_accounting (ticket_id)
 // ----------------------------------------------------------
 CREATE TABLE ticket_watcher (
     ticket_id BIGINT NOT NULL,
-    user_id BIGINT NOT NULL,
+    user_id INTEGER NOT NULL,
     create_time timestamp NOT NULL,
     create_by INTEGER NOT NULL,
     change_time timestamp NOT NULL,
